fix: timeline ternary and structure
