+++ src/proposals/components/timeline/static.ts
@@ -1,0 +1,15 @@
+import { styled } from '@mui/system';
+
+export const TimelineLineWrapper = styled('div')({
+  position: 'absolute',
+  left: 0,
+  height: 6,
+});
+
+export const TimelineCanceledItemWrapper = styled('div')({
+  display: 'flex',
+  position: 'absolute',
+  alignItems: 'center',
+  justifyContent: 'center',
+  top: 2,
+});
